Move tests outside source
