+++ packages/api/source/commands/database-reset.test.ts
@@ -8,7 +8,7 @@ import { ensureDirSync } from "fs-extra";
 import prompts from "prompts";
 import { dirSync, setGracefulCleanup } from "tmp";
 
-import { apiPackageJson } from "../test/fixtures";
+import { apiPackageJson } from "../../test/fixtures";
 import { Command } from "./database-reset";
 
 describe<{
