+++ packages/api/source/commands/env-get.test.ts
@@ -5,7 +5,7 @@ import { dirSync, setGracefulCleanup } from "tmp";
 
 import { Console } from "@mainsail/cli";
 import { describe } from "@mainsail/test-runner";
-import { apiPackageJson } from "../test/fixtures";
+import { apiPackageJson } from "../../test/fixtures";
 import { Command } from "./env-get";
 
 describe<{
